Only flush the input stream before reading data.
In the tools.rawInput() function, using the TCIOFLUSH queue selector
seemed to be discarding data from both the stdin and stdout queues, even
though the file descriptor was set to sys.stdin. This was giving me
intermittent missing output depending on the relative timing of the
stdout queue being pushed to the screen and tcflush() emptying it. For
example, sometimes on importing I would be asked "Is this correct?"
without seeing the proposed entry because it had been flushed.

This commit changes the tcflush() call to use TCIFLUSH to only empty the
input queue.